Change return value for getItem.
The HTML specification state that for non existent key, the getItem method must return null.
For now, the existing keys are kept in memory when setting a new one.
A simple lookup on that internal array will determine if a key is present.
Thus, as for now, a key is present only if it previously have been set.